refactor: replace custom form widgets with dedicated widget files for login and registration screens, improving code organization and readability
